feat: gate insufficient_evidence, cover all estimators, extract decision engine
Implements the recommendation/gating-engine issue cluster in one change.

#235 — Extract the decision layer into a focused `skdr_eval.recommendation`
module (Recommendation/RecommendationPolicy/Reason, GateResult/DiagnosticGate,
gate_diagnostics, _build_recommendation). Behaviour-preserving: every public
import path (`skdr_eval.*` and `skdr_eval.reporting.*`) is kept via re-export.

#196 — `_verdict_exit_code` now scans every estimator present in the artifact
(not just DR/SNDR), so a `do_not_deploy` from MRDR/SWITCH-DR/DRos/MIPS trips the
CI gate, and recommendation errors are logged instead of silently swallowed.

#197 — `insufficient_evidence` is a first-class, documented verdict with its own
CLI exit code (4), so an honest "we can't tell" cannot pass a CI gate as green.
do_not_deploy (3) takes precedence. Documented in README, metrics glossary, and
the report-interpretation guide.

#211 — New invariant tests for the decision layer: verdict-set completeness,
monotonicity (degrading a diagnostic never improves the verdict), and
card<->CLI exit-code consistency.

#245 — New regression tests pinning each documented CLI exit code (0/1/2/3/4).

Verification: ruff check, ruff format --check, mypy (0 issues), full pytest
suite (941 passed, 9 skipped).
